refactor based on project structure
