+++ internal/metrics/collector/otel_collector_plugin_test.go
@@ -2,7 +2,7 @@
 //
 // This source code is licensed under the Apache License, Version 2.0 license found in the
 // LICENSE file in the root directory of this source tree.
-package plugin
+package collector
 
 import (
 	"context"
